Layed the basis for better code organisation, but still a mess
