+++ src/config.hpp
@@ -13,7 +13,7 @@
  *	GNU General Public License for more details.
  *
  *	You should have received a copy of the GNU General Public License
- *	along with seriespl.  If not, see <http://www.gnu.org/licenses/>.
+ *	along with seriespl. If not, see <http://www.gnu.org/licenses/>.
  *
  ******************************************************************************/
